refactor: 패키지 구조 도메인 중심으로 변경
