--- src/test/java/usw/suwiki/BuildResultActionsException.java
@@ -1,7 +1,0 @@
-package usw.suwiki;
-
-public class BuildResultActionsException extends RuntimeException {
-    public BuildResultActionsException(Throwable cause) {
-        super(cause);
-    }
-}
