--- src/test/java/usw/suwiki/config/TestJpaConfig.java
@@ -1,20 +1,0 @@
-package usw.suwiki.config;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import javax.persistence.EntityManager;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
-
-@TestConfiguration
-@RequiredArgsConstructor
-public class TestJpaConfig {
-    @Autowired
-    private final EntityManager em;
-
-    @Bean
-    public JPAQueryFactory jpaQueryFactory() {
-        return new JPAQueryFactory(em);
-    }
-}
