+++ src/test/java/usw/suwiki/domain/blacklist/controller/BlacklistDomainControllerV2TestBase.java
@@ -1,0 +1,40 @@
+package usw.suwiki.domain.blacklist.controller;
+
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+import java.sql.Connection;
+import lombok.SneakyThrows;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.core.io.ClassPathResource;
+import org.springframework.jdbc.datasource.init.ScriptUtils;
+import usw.suwiki.global.IntegrationTestBase;
+import usw.suwiki.global.jwt.JwtAgent;
+
+
+public class BlacklistDomainControllerV2TestBase extends IntegrationTestBase {
+
+    @MockBean
+    JwtAgent jwtAgent;
+
+    @BeforeAll
+    public void init() throws Exception {
+        try (Connection conn = dataSource.getConnection()) {
+            ScriptUtils.executeSqlScript(conn, new ClassPathResource("/data/insert-user.sql"));
+            ScriptUtils.executeSqlScript(conn, new ClassPathResource("/data/insert-confirmationtoken.sql"));
+            ScriptUtils.executeSqlScript(conn, new ClassPathResource("/data/insert-lecture.sql"));
+            ScriptUtils.executeSqlScript(conn, new ClassPathResource("/data/insert-exampost.sql"));
+            ScriptUtils.executeSqlScript(conn, new ClassPathResource("/data/insert-viewexam.sql"));
+        }
+    }
+
+    @SneakyThrows(Exception.class)
+    @Test
+    @DisplayName("블랙리스트 사유 불러오기")
+    void 블랙리스트_사유_불러오기() {
+        executeGetRequestWithAuthorizationResultActions("/v2/blacklist/logs")
+                .andExpect(status().isOk());
+    }
+}
