+++ src/test/java/usw/suwiki/domain/lecture/service/InsertNewLectureTest.java
