+++ src/test/java/usw/suwiki/domain/lecture/template/LectureDetailTemplate.java
@@ -1,0 +1,41 @@
+package usw.suwiki.domain.lecture.template;
+
+import usw.suwiki.domain.lecture.domain.LectureDetail;
+
+public class LectureDetailTemplate {
+    public static final String PLACE_SCHEDULE = "IT307(월6,7,8)";
+    public static final String CODE = "01760";
+    public static final double POINT = 0;
+    public static final String CAPPR_TYPE = "A형(강의식 수업)";
+    public static final String DICL_NO = "001";
+    public static final int GRADE = 2;
+    public static final String EVALUATE_TYPE = "상대평가";
+
+    public static LectureDetail createFirstDummy() {
+        return createDummy(PLACE_SCHEDULE, CODE, POINT, CAPPR_TYPE, DICL_NO, GRADE, EVALUATE_TYPE);
+    }
+
+    public static LectureDetail varyPlaceSchedule(String placeSchedule) {
+        return createDummy(placeSchedule, CODE, POINT, CAPPR_TYPE, DICL_NO, GRADE, EVALUATE_TYPE);
+    }
+
+    public static LectureDetail createDummy(
+            String placeSchedule,
+            String code,
+            double point,
+            String capprType,
+            String diclNo,
+            int grade,
+            String evaluateType
+    ) {
+        return LectureDetail.builder()
+                .placeSchedule(placeSchedule)
+                .code(code)
+                .point(point)
+                .capprType(capprType)
+                .diclNo(diclNo)
+                .grade(grade)
+                .evaluateType(evaluateType)
+                .build();
+    }
+}
