+++ src/test/java/usw/suwiki/global/exception/BuildResultActionsException.java
@@ -1,0 +1,7 @@
+package usw.suwiki.global.exception;
+
+public class BuildResultActionsException extends RuntimeException {
+    public BuildResultActionsException(Throwable cause) {
+        super(cause);
+    }
+}
